Refactor Homepage codec
