+++ .changeset/mighty-games-drum.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": patch
+---
+
+Refactor homepage codec
