+++ apps/nextjs-website/src/app/page.tsx
@@ -12,6 +12,63 @@ import { getHomepageProps } from '@/lib/cmsApi';
 import BlocksRendererClient from '@/components/molecules/BlocksRendererClient/BlocksRendererClient';
 import Ecosystem from '@/components/organisms/Ecosystem/Ecosystem';
 import ContentWrapper from '@/components/atoms/ContentWrapper/ContentWrapper';
+import { Media } from '@/lib/strapi/codecs/MediaCodec';
+import { CardsGridProps } from '@/components/molecules/CardsGrid/CardsGrid';
+import { CtaSlideProps } from '@/components/atoms/CtaSlide/CtaSlide';
+import { Webinar } from '@/lib/types/webinar';
+import { SEO } from '@/lib/types/seo';
+
+type NewsShowcaseItemProps = {
+  readonly comingSoon?: boolean;
+  readonly title: string;
+  readonly publishedAt?: Date;
+  readonly link: {
+    readonly text: string;
+    readonly url: string;
+    readonly target?: '_self' | '_blank' | '_parent' | '_top';
+  };
+  readonly image?: Media;
+};
+
+type NewsShowcaseProps = {
+  readonly title: string;
+  readonly items: readonly NewsShowcaseItemProps[];
+};
+
+type EcosystemSolutionsCtaProps = {
+  readonly variant?: 'text' | 'contained' | 'outlined';
+  readonly link: {
+    readonly href: string;
+    readonly text: string;
+    readonly target?: '_self' | '_blank' | '_parent' | '_top';
+  };
+};
+
+type EcosystemProps = {
+  readonly title: string;
+  readonly productsTabName: string;
+  readonly products: CardsGridProps['cards'];
+  readonly solutionsTabName: string;
+  readonly solutions?: CardsGridProps['cards'];
+  readonly solutionsCta?: EcosystemSolutionsCtaProps;
+};
+
+type ComingSoonDocumentationProps = {
+  readonly title: string;
+  readonly links: readonly {
+    readonly text: string;
+    readonly href: string;
+  }[];
+};
+
+export type HomepageProps = {
+  readonly hero: readonly CtaSlideProps[];
+  readonly newsShowcase: NewsShowcaseProps;
+  readonly ecosystem: EcosystemProps;
+  readonly webinars: readonly Webinar[];
+  readonly comingsoonDocumentation: ComingSoonDocumentationProps;
+  readonly seo?: SEO;
+};
 
 export async function generateMetadata(): Promise<Metadata> {
   const homepage = await getHomepageProps();
@@ -37,7 +94,7 @@ const NotSsrWebinarsSection = dynamic(
 );
 
 const Home = async () => {
-  const homepage = await getHomepageProps();
+  const homepage: HomepageProps = await getHomepageProps();
 
   return (
     <ContentWrapper>
