+++ apps/nextjs-website/src/lib/strapi/__tests__/fixtures/media.ts
@@ -22,75 +22,77 @@ export const mediaVectorJson = {
   },
 };
 
+export const mediaRasterAttributesJson = {
+  name: 'hero.jpg',
+  alternativeText: null,
+  caption: null,
+  width: 1440,
+  height: 720,
+  formats: {
+    medium: {
+      name: 'medium_hero.jpg',
+      hash: 'medium_hero_b2b3247d4c',
+      ext: '.jpg',
+      mime: 'image/jpeg',
+      path: null,
+      width: 750,
+      height: 375,
+      size: 3.53,
+      sizeInBytes: 3527,
+      url: 'http://localhost:1337/uploads/medium_hero_b2b3247d4c.jpg',
+    },
+    large: {
+      name: 'large_hero.jpg',
+      hash: 'large_hero_b2b3247d4c',
+      ext: '.jpg',
+      mime: 'image/jpeg',
+      path: null,
+      width: 1000,
+      height: 500,
+      size: 5.42,
+      sizeInBytes: 5424,
+      url: 'http://localhost:1337/uploads/large_hero_b2b3247d4c.jpg',
+    },
+    small: {
+      name: 'small_hero.jpg',
+      hash: 'small_hero_b2b3247d4c',
+      ext: '.jpg',
+      mime: 'image/jpeg',
+      path: null,
+      width: 500,
+      height: 250,
+      size: 2.08,
+      sizeInBytes: 2081,
+      url: 'http://localhost:1337/uploads/small_hero_b2b3247d4c.jpg',
+    },
+    thumbnail: {
+      name: 'thumbnail_hero.jpg',
+      hash: 'thumbnail_hero_b2b3247d4c',
+      ext: '.jpg',
+      mime: 'image/jpeg',
+      path: null,
+      width: 245,
+      height: 123,
+      size: 0.98,
+      sizeInBytes: 975,
+      url: 'http://localhost:1337/uploads/thumbnail_hero_b2b3247d4c.jpg',
+    },
+  },
+  hash: 'hero_b2b3247d4c',
+  ext: '.jpg',
+  mime: 'image/jpeg',
+  size: 51.34,
+  url: 'http://localhost:1337/uploads/hero_b2b3247d4c.jpg',
+  previewUrl: null,
+  provider: 'strapi-provider-upload-custom',
+  provider_metadata: null,
+  createdAt: '2024-08-21T13:21:14.822Z',
+  updatedAt: '2024-08-21T13:21:14.822Z',
+};
+
 export const mediaRasterJson = {
   data: {
     id: 100,
-    attributes: {
-      name: 'hero.jpg',
-      alternativeText: null,
-      caption: null,
-      width: 1440,
-      height: 720,
-      formats: {
-        medium: {
-          name: 'medium_hero.jpg',
-          hash: 'medium_hero_b2b3247d4c',
-          ext: '.jpg',
-          mime: 'image/jpeg',
-          path: null,
-          width: 750,
-          height: 375,
-          size: 3.53,
-          sizeInBytes: 3527,
-          url: 'http://localhost:1337/uploads/medium_hero_b2b3247d4c.jpg',
-        },
-        large: {
-          name: 'large_hero.jpg',
-          hash: 'large_hero_b2b3247d4c',
-          ext: '.jpg',
-          mime: 'image/jpeg',
-          path: null,
-          width: 1000,
-          height: 500,
-          size: 5.42,
-          sizeInBytes: 5424,
-          url: 'http://localhost:1337/uploads/large_hero_b2b3247d4c.jpg',
-        },
-        small: {
-          name: 'small_hero.jpg',
-          hash: 'small_hero_b2b3247d4c',
-          ext: '.jpg',
-          mime: 'image/jpeg',
-          path: null,
-          width: 500,
-          height: 250,
-          size: 2.08,
-          sizeInBytes: 2081,
-          url: 'http://localhost:1337/uploads/small_hero_b2b3247d4c.jpg',
-        },
-        thumbnail: {
-          name: 'thumbnail_hero.jpg',
-          hash: 'thumbnail_hero_b2b3247d4c',
-          ext: '.jpg',
-          mime: 'image/jpeg',
-          path: null,
-          width: 245,
-          height: 123,
-          size: 0.98,
-          sizeInBytes: 975,
-          url: 'http://localhost:1337/uploads/thumbnail_hero_b2b3247d4c.jpg',
-        },
-      },
-      hash: 'hero_b2b3247d4c',
-      ext: '.jpg',
-      mime: 'image/jpeg',
-      size: 51.34,
-      url: 'http://localhost:1337/uploads/hero_b2b3247d4c.jpg',
-      previewUrl: null,
-      provider: 'strapi-provider-upload-custom',
-      provider_metadata: null,
-      createdAt: '2024-08-21T13:21:14.822Z',
-      updatedAt: '2024-08-21T13:21:14.822Z',
-    },
+    attributes: mediaRasterAttributesJson,
   },
 };
